+++ src/crypt/trivium.c
@@ -122,7 +122,7 @@ void cry_trivium_iv_set(cry_trivium_ctx *ctx, const unsigned char *iv,
 
 
 static void operate(cry_trivium_ctx *ctx, unsigned char *dst,
-                    const unsigned char *src, unsigned int size)
+                    const unsigned char *src, size_t size)
 
 {
     size_t i;
