+++ test/test.h
@@ -23,7 +23,7 @@ void func_test(const char *datafile, dispatch_func_t dispatch);
 
 void asc_to_raw(const char *asc, size_t size, unsigned char *raw);
 
-int raw_init(unsigned char *raw, unsigned int rawlen, const char *asc);
+int raw_init(unsigned char *raw, size_t rawlen, const char *asc);
 
 
 #define ARLEN(ar) (sizeof(ar)/sizeof(ar[0]))
